Only removing apps during PR events

+++ Actions/RunPipeline/RunPipeline.ps1
@@ -508,21 +508,23 @@ try {
         -appBuild $appBuild -appRevision $appRevision `
         -uninstallRemovedApps
 
-    # If any apps were downloaded as part of incremental builds, we should remove them again after the build
-    $downloadedAppsByType | ForEach-Object {
-        if ($_.downloadedApps) {
-            $mask = $_.mask
-            $thisArtifactFolder = Join-Path $buildArtifactFolder $mask
-            Write-Host "Removing downloaded apps from $thisArtifactFolder"
-            foreach($downloadedApp in $_.downloadedApps) {
-                $thisApp = Join-Path $thisArtifactFolder $downloadedApp
-                Write-Host "Removing: $thisApp"
-                if (Test-Path $thisApp) {
-                    Remove-Item $thisApp
+    # If any apps were downloaded as part of incremental builds in a pr, we should remove them again after the build
+    if ($ENV:GITHUB_EVENT_NAME -like 'pull_request*') {
+        $downloadedAppsByType | ForEach-Object {
+            if ($_.downloadedApps) {
+                $mask = $_.mask
+                $thisArtifactFolder = Join-Path $buildArtifactFolder $mask
+                Write-Host "Removing downloaded apps from $thisArtifactFolder"
+                foreach($downloadedApp in $_.downloadedApps) {
+                    $thisApp = Join-Path $thisArtifactFolder $downloadedApp
+                    Write-Host "Removing: $thisApp"
+                    if (Test-Path $thisApp) {
+                        Remove-Item $thisApp
+                    }
                 }
             }
         }
-    }
+    } 
 
     if ($containerBaseFolder) {
         Write-Host "Copy artifacts and build output back from build container"
